Farmer's Wallet Final/Farmer's Wallet Final/code_1/requirements-clean.txt
Flask==3.1.0
python-dotenv==1.1.0
google-generativeai==0.8.5
joblib==1.5.0
scikit-learn==1.5.0
pandas==2.0.3
numpy==1.26.0
markdown==3.5.1
xgboost==2.0.3
imbalanced-learn==0.11.0
